+++ rinokeras/core/v1x/common/layers/masking.py
@@ -9,9 +9,11 @@ from tensorflow.keras.layers import Layer  # pylint: disable=F0401
 import tensorflow.keras.backend as K  # pylint: disable=F0401
 
 
-class MaskInput(Layer):
+class BERTRandomReplaceMask(Layer):
     """
-    Replaces some percentage of the input with a mask token. Used for implementing BERT style models.
+    Replaces some percentage of the input with a mask token. Used for
+    mplementing BERT style models. This is actually slightly more complex - it 
+    does one of three things
 
     Based on https://arxiv.org/abs/1810.04805.
 
@@ -20,19 +22,27 @@ class MaskInput(Layer):
         mask_token (int): Token to replace masked input with
     """
 
-    def __init__(self, percentage: float, mask_token: int, n_symbols: Optional[int] = None, *args, **kwargs) -> None:
+    def __init__(self,
+                 percentage: float,
+                 mask_token: int,
+                 n_symbols: Optional[int] = None,
+                 *args, **kwargs) -> None:
         super().__init__(*args, **kwargs)
         if not 0 <= percentage < 1:
-            raise ValueError("Masking percentage must be in [0, 1). Received {}".format(percentage))
+            raise ValueError("Masking percentage must be in [0, 1).\
+                Received {}".format(percentage))
         self.percentage = percentage
         self.mask_token = mask_token
         self.n_symbols = n_symbols
 
-    def call(self, inputs: tf.Tensor, valid_mask: Optional[tf.Tensor] = None, n_symbols: Optional[int] = None):
+    def call(self,
+             inputs: tf.Tensor,
+             mask: Optional[tf.Tensor] = None,
+             n_symbols: Optional[int] = None):
         """
         Args:
             inputs (tf.Tensor[ndims=2, int]): Tensor of values to mask
-            valid_mask (Optional[tf.Tensor[bool]]): Locations in the inputs to that are valid
+            mask (Optional[tf.Tensor[bool]]): Locations in the inputs to that are valid
                                                      (i.e. not padding, start tokens, etc.)
         Returns:
             masked_inputs (tf.Tensor[ndims=2, int]): Tensor of masked values
@@ -47,16 +57,18 @@ class MaskInput(Layer):
 
         bert_mask = K.random_uniform(mask_shape) < self.percentage
 
-        if valid_mask is not None:
-            bert_mask &= valid_mask
+        if mask is not None:
+            bert_mask &= mask
 
         if not discrete:
             bert_mask = tf.expand_dims(bert_mask, -1)
 
-        masked_inputs = inputs * tf.cast(~bert_mask, inputs.dtype)  # type: ignore
+        masked_inputs = inputs * \
+            tf.cast(~bert_mask, inputs.dtype)  # type: ignore
 
         token_bert_mask = K.random_uniform(K.shape(bert_mask)) < 0.8
-        random_bert_mask = (K.random_uniform(K.shape(bert_mask)) < 0.1) & ~token_bert_mask
+        random_bert_mask = (K.random_uniform(
+            K.shape(bert_mask)) < 0.1) & ~token_bert_mask
         true_bert_mask = ~token_bert_mask & ~random_bert_mask
 
         token_bert_mask = tf.cast(token_bert_mask & bert_mask, inputs.dtype)
@@ -70,7 +82,8 @@ class MaskInput(Layer):
             masked_inputs += K.random_uniform(
                 K.shape(bert_mask), 0, n_symbols, dtype=inputs.dtype) * random_bert_mask
         else:
-            masked_inputs += (K.random_normal(K.shape(masked_inputs)) + inputs) * random_bert_mask
+            masked_inputs += (K.random_normal(K.shape(masked_inputs)
+                                              ) + inputs) * random_bert_mask
 
         masked_inputs += inputs * true_bert_mask