+++ rinokeras/testing/v1x/test_losses.py
@@ -1,0 +1,12 @@
+"""
+Loss testing
+"""
+import numpy as np
+import tensorflow as tf
+import warnings
+import tempfile
+
+from rinokeras.testing.utils import *
+
+def test_losses():
+    pass