+++ rinokeras/testing/v1x/test_optimizers.py
@@ -1,0 +1,12 @@
+"""
+Optimizer testing
+"""
+import numpy as np
+import tensorflow as tf
+import warnings
+import tempfile
+
+from rinokeras.testing.utils import *
+
+def test_optimizers():
+    pass