+++ rinokeras/testing/v1x/test_rnn.py
@@ -1,0 +1,12 @@
+"""
+RNN testing
+"""
+import numpy as np
+import tensorflow as tf
+import warnings
+import tempfile
+
+from rinokeras.testing.utils import *
+
+def test_rnn():
+    pass